+++ packages/react/src/components/radio-group/examples/initial-value.tsx
@@ -10,6 +10,7 @@ export const InitialValue = () => {
         <RadioGroup.Item key={framework} value={framework}>
           <RadioGroup.ItemText>{framework}</RadioGroup.ItemText>
           <RadioGroup.ItemControl />
+          <RadioGroup.ItemHiddenInput />
         </RadioGroup.Item>
       ))}
     </RadioGroup.Root>