+++ packages/react/src/components/radio-group/index.ts
@@ -4,6 +4,10 @@ export { RadioGroupIndicator, type RadioGroupIndicatorProps } from './radio-grou
 export { RadioGroupItem, type RadioGroupItemProps } from './radio-group-item'
 export { RadioGroupItemContext, type RadioGroupItemContextProps } from './radio-group-item-context'
 export { RadioGroupItemControl, type RadioGroupItemControlProps } from './radio-group-item-control'
+export {
+  RadioGroupItemHiddenInput,
+  type RadioGroupItemHiddenInputProps,
+} from './radio-group-item-hidden-input'
 export { RadioGroupItemText, type RadioGroupItemTextProps } from './radio-group-item-text'
 export { RadioGroupLabel, type RadioGroupLabelProps } from './radio-group-label'
 export { RadioGroupRoot, type RadioGroupRootProps } from './radio-group-root'