+++ packages/solid/src/components/radio-group/examples/basic.tsx
@@ -13,6 +13,7 @@ export const Basic = () => {
           <RadioGroup.Item value={framework()}>
             <RadioGroup.ItemText>{framework()}</RadioGroup.ItemText>
             <RadioGroup.ItemControl />
+            <RadioGroup.ItemHiddenInput />
           </RadioGroup.Item>
         )}
       </Index>